refactor: move CampaignData to repositories
